Tie Up Loose Ends with fixes.

+++ wp-content/themes/block-theme/functions.php
@@ -6,26 +6,26 @@ if ( ! function_exists( 'eri_scaffold_support' ) ) :
 		if ( ! 'eri-scaffold-block-theme' === wp_get_theme()->get( 'TextDomain' ) ) {
 			load_theme_textdomain( wp_get_theme()->get( 'TextDomain' ) );
 		}
-
+		
 		// Alignwide and alignfull classes in the block editor.
 		add_theme_support( 'align-wide' );
-
+		
 		// Add support for link color control.
 		add_theme_support( 'link-color' );
-
+		
 		// Add support for responsive embedded content.
 		// https://github.com/WordPress/gutenberg/issues/26901
 		add_theme_support( 'responsive-embeds' );
-
+		
 		// Add support for editor styles.
 		add_theme_support( 'editor-styles' );
-
+		
 		// Add support for post thumbnails.
 		add_theme_support( 'post-thumbnails' );
-
+		
 		// Experimental support for adding blocks inside nav menus
 		add_theme_support( 'block-nav-menus' );
-
+		
 		// Enqueue editor styles.
 		add_editor_style(
 			array(
@@ -33,7 +33,7 @@ if ( ! function_exists( 'eri_scaffold_support' ) ) :
 			)
 		);
 
-		// Register two nav menus if Gutenberg is activated (otherwise the __experimentalMenuLocation attribute isn't available)
+			// Register two nav menus if Gutenberg is activated (otherwise the __experimentalMenuLocation attribute isn't available)
 		if ( defined( 'IS_GUTENBERG_PLUGIN' ) ) {
 			register_nav_menus(
 				array(
@@ -42,7 +42,7 @@ if ( ! function_exists( 'eri_scaffold_support' ) ) :
 				)
 			);
 		}
-
+			
 		add_filter(
 			'block_editor_settings_all',
 			function( $settings ) {
@@ -50,7 +50,7 @@ if ( ! function_exists( 'eri_scaffold_support' ) ) :
 				return $settings;
 			}
 		);
-
+		
 		// Add support for core custom logo.
 		add_theme_support(
 			'custom-logo',
@@ -61,10 +61,10 @@ if ( ! function_exists( 'eri_scaffold_support' ) ) :
 				'flex-height' => true,
 			)
 		);
-
+			
 	}
+	add_action( 'after_setup_theme', 'eri_scaffold_support', 9 );
 endif;
-add_action( 'after_setup_theme', 'eri_scaffold_support', 9 );
 
 /**
  *
@@ -77,7 +77,7 @@ function eri_scaffold_editor_styles() {
 			'/assets/theme.css'
 		);
 	}
-
+	
 	wp_enqueue_style( 'eri-scaffold-admin-styles', get_template_directory_uri() . '/assets/admin.css', array(), wp_get_theme()->get( 'Version' ) );
 }
 add_action( 'admin_init', 'eri_scaffold_editor_styles' );
@@ -91,12 +91,12 @@ function eri_scaffold_scripts() {
 	if ( file_exists( get_stylesheet_directory() . '/assets/ponyfill.css' ) ) {
 		wp_enqueue_style( 'eri-scaffold-ponyfill', get_template_directory_uri() . '/assets/ponyfill.css', array(), filemtime( get_stylesheet_directory() . '/assets/ponyfill.css' ) );
 	}
-
+	
 	// Add the child theme CSS if it exists.
 	if ( file_exists( get_stylesheet_directory() . '/assets/theme.css' ) ) {
 		wp_enqueue_style( 'eri-scaffold-child-styles', get_stylesheet_directory_uri() . '/assets/theme.css', array( 'eri-scaffold-ponyfill' ), filemtime( get_stylesheet_directory() . '/assets/theme.css' ) );
 	}
-
+	
 	if ( file_exists( get_stylesheet_directory() . '/assets/scripts.js' ) ) {
 		wp_enqueue_script( 'eri-scaffold-scripts', get_stylesheet_directory_uri() . '/assets/scripts.js', array(), filemtime( get_stylesheet_directory() . '/assets/scripts.js' ) );
 	}
@@ -114,16 +114,16 @@ add_action( 'wp_enqueue_scripts', 'eri_scaffold_scripts' );
  * @return void
  */
 function eri_scaffold_editor_scripts() {
-
-    wp_register_script(
-        'editor-scripts',
+	
+	wp_register_script(
+		'editor-scripts',
         get_stylesheet_directory_uri() . '/assets/js/editor-scripts.js',
         array( 'wp-blocks', 'wp-i18n', 'wp-element', 'wp-components', 'wp-editor' ),
         filemtime( trailingslashit( get_template_directory() ) . 'assets/js/editor-scripts.js' ),
         array( 'in_footer' => true )
     );
     wp_enqueue_script('editor-scripts');
-
+	
 }
 add_action( 'enqueue_block_editor_assets', 'eri_scaffold_editor_scripts' );
 
@@ -173,12 +173,12 @@ if ( file_exists( get_stylesheet_directory() . '/inc/block-patterns.php' ) ) {
  * @return mixed
  */
 function eri_scaffold_body_class( $classes ) {
-
-    if ( ! is_admin() )
-
+	
+	if ( ! is_admin() )
+	
     $classes[] = 'eri-scaffold-not-admin';
     return $classes;
-
+	
 }
 add_filter( 'body_class', 'eri_scaffold_body_class' );
 
@@ -188,7 +188,7 @@ add_filter( 'body_class', 'eri_scaffold_body_class' );
  * @return void
  */
 function eri_scaffold_wp_body_open() {
-    echo '<a class="skip-to-main" href="#main">Skip to main content</a>';
+	echo '<a class="skip-to-main" href="#main">Skip to main content</a>';
 }
 add_action( 'wp_body_open', 'eri_scaffold_wp_body_open' );
 
@@ -200,10 +200,10 @@ register_nav_menus( array( 'actions' => esc_html__( 'Actions', 'eri-scaffold-blo
 register_nav_menus( array( 'bottom_bar' => esc_html__( 'Bottom Bar', 'eri-scaffold-block-theme' ) ) );
 
 if( function_exists('acf_add_options_page') ) {
-
-    // Add Site Settings options page
+	
+	// Add Site Settings options page
     acf_add_options_page(array(
-        'page_title' 	=> 'Site Settings',
+		'page_title' 	=> 'Site Settings',
         'menu_title'	=> 'Site Settings',
         'menu_slug' 	=> 'eri-scaffold-settings',
         'capability'	=> 'edit_posts',
@@ -217,14 +217,14 @@ See:
 https://app.asana.com/0/1205013361806811/1206339105193214/f
 */
 add_filter( 'rank_math/opengraph/content_image_cache', function( $ret ) {
-    return false;
+	return false;
 }, PHP_INT_MAX );
 
 /*
 Turns off the Private: prefix for private post titles
 */
 function eri_scaffold_private_title_format() {
-    return '%s';
+	return '%s';
 }
 add_filter('private_title_format','eri_scaffold_private_title_format');
 add_filter('protected_title_format', 'eri_scaffold_private_title_format');
@@ -260,16 +260,16 @@ add_filter( 'image_editor_output_format', 'ngi_filter_image_editor_output_format
  * @since 1.0.0
  */
 function eri_scaffold_enqueue_block_styles() {
-    // If you want to add styles exclusively for the block editor, you can use:
+	// If you want to add styles exclusively for the block editor, you can use:
     if ( is_admin() ) {
-        // Register the editor stylesheet located in the theme's root directory.
+		// Register the editor stylesheet located in the theme's root directory.
         wp_register_style(
-            'blocks-admin',
+			'blocks-admin',
             get_theme_file_uri('assets/admin.css'),
             [],
             '1.0'
         );
-
+		
         // Enqueue editor styles.
         wp_enqueue_style( 'blocks-admin' );
     }
@@ -278,23 +278,23 @@ function eri_scaffold_enqueue_block_styles() {
 add_action( 'enqueue_block_assets', 'eri_scaffold_enqueue_block_styles' );
 
 function add_all_category_slugs_to_post_class( $classes, $class, $post_id ) {
-    if ( 'post' === get_post_type() ) {
-        $categories = wp_get_post_terms( $post_id, 'category' );
-
+	if ( 'post' === get_post_type() ) {
+		$categories = wp_get_post_terms( $post_id, 'category' );
+		
         foreach ( $categories as $category ) {
-            $classes[] = 'category-' . $category->slug;
+			$classes[] = 'category-' . $category->slug;
         }
     }
-
+	
 	if ( 'profiles' === get_post_type() ) {
 		$post_id = get_the_ID();
 		$categories = wp_get_post_terms( $post_id, 'profile-categories' );
-
+		
 		foreach ( $categories as $category ) {
 			$classes[] = 'profile-categories-' . $category->slug;
 		}
 	}
-
+	
     return $classes;
 }
 add_filter( 'post_class', 'add_all_category_slugs_to_post_class', 10, 3 );
@@ -303,25 +303,25 @@ function add_all_category_slugs_to_body_class( $classes ) {
 	if ( is_front_page() ) {
 		return $classes;
 	}
-
+	
 	if ( 'post' === get_post_type() ) {
 		$post_id = get_the_ID();
         $categories = wp_get_post_terms( $post_id, 'category' );
 
         foreach ( $categories as $category ) {
-            $classes[] = 'category-' . $category->slug;
+			$classes[] = 'category-' . $category->slug;
         }
     }
-
+	
 	if ( 'profiles' === get_post_type() ) {
 		$post_id = get_the_ID();
 		$categories = wp_get_post_terms( $post_id, 'profile-categories' );
-
+		
 		foreach ( $categories as $category ) {
 			$classes[] = 'profile-categories-' . $category->slug;
 		}
 	}
-
+	
     return $classes;
 }
 add_filter( 'body_class', 'add_all_category_slugs_to_body_class' );
@@ -330,49 +330,53 @@ function add_all_category_slugs_to_admin_body_class( $classes ) {
 	if ( 'post' === get_post_type() ) {
 		$post_id = get_the_ID();
         $categories = wp_get_post_terms( $post_id, 'category' );
-
+		
         foreach ( $categories as $category ) {
-            $classes .= ' category-' . $category->slug;
+			$classes .= ' category-' . $category->slug;
         }
     }
-
+	
 	if ( 'profiles' === get_post_type() ) {
 		$post_id = get_the_ID();
 		$categories = wp_get_post_terms( $post_id, 'profile-categories' );
-
+		
 		foreach ( $categories as $category ) {
 			$classes .= ' profile-categories-' . $category->slug;
 		}
 	}
-
+	
     return $classes;
 }
 add_filter( 'admin_body_class', 'add_all_category_slugs_to_admin_body_class' );
 
 function add_category_styles_to_head() {
-    // Retrieve all categories
+	// Retrieve all categories
     $categories = get_categories();
-
+	
     if ( empty( $categories ) ) {
-        return;
+		return;
     }
 
+	if ( ! function_exists( 'get_field' ) ) {
+		return;
+	}
+	
 	$profile_categories = get_terms( array(
 		'taxonomy' => 'profile-categories',
 		'hide_empty' => true,
-	) );
-
-    // Start output buffering
-    ob_start();
-    echo '<style>';
-    foreach ( $categories as $category ) {
-		$color = get_field( 'category_color', 'term_' . $category->term_id );
-        echo '.category-' . esc_html( $category->slug ) . ', .category-' . esc_html( $category->slug ) . ' .editor-styles-wrapper { --eri-scaffold--color--accent: ' . esc_html( $color ) . '; }';
-    }
-
-	foreach ( $profile_categories as $category ) {
-		$color = get_field( 'category_color', 'term_' . $category->term_id );
-		echo '.profile-categories-' . esc_html( $category->slug ) . ', .profile-categories-' . esc_html( $category->slug ) . ' .editor-styles-wrapper { --eri-scaffold--color--accent: ' . esc_html( $color ) . '; }';
+		) );
+		
+		// Start output buffering
+		ob_start();
+		echo '<style>';
+		foreach ( $categories as $category ) {
+			$color = get_field( 'category_color', 'term_' . $category->term_id );
+			echo '.category-' . esc_html( $category->slug ) . ', .category-' . esc_html( $category->slug ) . ' .editor-styles-wrapper { --eri-scaffold--color--accent: ' . esc_html( $color ) . '; }';
+		}
+		
+		foreach ( $profile_categories as $category ) {
+			$color = get_field( 'category_color', 'term_' . $category->term_id );
+			echo '.profile-categories-' . esc_html( $category->slug ) . ', .profile-categories-' . esc_html( $category->slug ) . ' .editor-styles-wrapper { --eri-scaffold--color--accent: ' . esc_html( $color ) . '; }';
 	}
     echo '</style>';
     // Output and clear buffer
@@ -384,41 +388,45 @@ add_action( 'admin_head', 'add_category_styles_to_head', 100 );
 function add_category_styles_to_admin_head() {
 	$color = null;
 
+	if ( ! function_exists( 'get_field' ) ) {
+		return;
+	}
+
 	// Retrieve all categories
     if ( 'post' === get_post_type() ) {
 		$post_id = get_the_ID();
         $categories = wp_get_post_terms( $post_id, 'category' );
 		$category = null;
-
+		
         foreach ( $categories as $cat ) {
-            $category = $cat;
+			$category = $cat;
         }
-
+		
 		if ( ! empty( $category ) ) {
 			$color = get_field( 'category_color', 'term_' . $category->term_id );
 		}
     }
-
+	
 	if ( 'profile' === get_post_type() ) {
 		$post_id = get_the_ID();
         $categories = wp_get_post_terms( $post_id, 'profile-categories' );
 		$category = null;
-
+		
         foreach ( $categories as $cat ) {
-            $category = $cat;
+			$category = $cat;
         }
-
+		
 		if ( ! empty( $category ) ) {
 			$color = get_field( 'category_color', 'term_' . $category->term_id );
 		}
     }
-
+	
     ?>
 	<style><?php
 		if ( ! empty( $color ) ) {
 			echo 'body { --eri-scaffold--color--accent: ' . esc_html( $color ) . '; }';
 		}
-    ?></style>
+		?></style>
 	<?php
 }
 add_action( 'admin_head', 'add_category_styles_to_admin_head', 1000000 );
@@ -428,5 +436,5 @@ add_filter( 'gform_disable_css', '__return_true' );
 
 // Reduce Wordpress default excerpt length
 add_filter( 'excerpt_length', function( $length ) {
-    return 30;
+	return 30;
 }, PHP_INT_MAX);
